requirements.txt
numpy
matplotlib
Pillow
torch
torchvision
scikit-learn
tqdm